+++ Team_Project_Paint/Team_Project_Paint/Enum/NameForShapeFactory.cs
@@ -13,6 +13,7 @@ namespace Team_Project_Paint.Enum
         Ellipse,
         Curve,
         Line,
+        Triangle,
         Select
     }
